requirements.txt
PyQt6>=6.7.1
torch>=2.1.0
Pillow>=10.1.0
opencv-python>=4.10.0.84
lama-cleaner>=1.2.5
flask
